app/requirements.txt
aiofiles==22.1.0
aiolimiter==1.0.0
anyio==3.6.2
APScheduler==3.9.1.post1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
pycparser==2.21
python-telegram-bot==20.0a6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
tornado==6.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
